requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.5
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
binaryornot==0.4.4
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==43.0.3
decorator==5.1.1
dill==0.3.9
Django==5.1.2
django-appconf==1.0.6
django-browser-reload==1.17.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-crontab==0.7.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-phonenumber-field==8.0.0
django-tailwind==3.8.0
django-timezone-field==7.0
django-unicorn==0.61.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-redesign==0.4.0
environ==1.0
Faker==30.8.2
gevent==24.10.3
google-api-core==2.22.0
google-api-python-client==2.127.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
kombu==5.4.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mercadopago==2.2.3
ndg-httpsclient==0.5.1
oauthlib==3.2.2
orjson==3.10.11
packaging==24.1
phonenumbers==8.13.48
pika==0.13.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
PyAMQP==0.1.0.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==1.7.1
pylint==3.3.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==8.3.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
rcssmin==1.1.3
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==13.9.4
rjsmin==1.2.3
rsa==4.9
setuptools==75.3.0
shortuuid==1.0.13
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.0.2
tomlkit==0.13.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.1
zope.event==5.0
zope.interface==7.1.1
